+++ spec/rails_app/app/components/grid/crud_inline.rb
@@ -1,0 +1,6 @@
+class Grid::CrudInline < Grid::Crud
+  def configure(c)
+    super
+    c.edit_inline = true
+  end
+end