+++ spec/rails_app/app/components/grid/custom_primary_key.rb
@@ -3,6 +3,7 @@ class Grid::CustomPrimaryKey < Netzke::Basepack::Grid
     c.title = "Books (model with custom primary key)"
     c.model = "BookWithCustomPrimaryKey"
     c.columns = [:author__name, :title]
+    c.edit_inline = true
     super
   end
 end