+++ spec/rails_app/app/components/grid/inline_editing.rb
@@ -1,0 +1,7 @@
+class Grid::InlineEditing < Netzke::Basepack::Grid
+  def configure(c)
+    super
+    c.model = 'Author'
+    c.edit_inline = true
+  end
+end