+++ spec/rails_app/app/example_components/grid_with_disabled_editing.rb
@@ -2,7 +2,7 @@ class GridWithDisabledEditing < Netzke::Basepack::Grid
   def configure(c)
     super
     c.model = 'Book'
-    c.enable_edit_inline = false
-    c.enable_edit_in_form = false
+    c.edit_inline = false
+    c.edit_in_form = false
   end
 end