+++ spec/rails_app/app/example_components/grid_with_in_form_editing_only.rb
@@ -6,6 +6,6 @@ class GridWithInFormEditingOnly < Netzke::Basepack::Grid
   def configure(c)
     super
     c.model = 'Book'
-    c.enable_edit_inline = false
+    c.edit_inline = false
   end
 end