+++ spec/rails_app/app/example_components/grid_with_inline_editing_only.rb
@@ -2,6 +2,6 @@ class GridWithInlineEditingOnly < Netzke::Basepack::Grid
   def configure(c)
     super
     c.model = 'Book'
-    c.enable_edit_in_form = false
+    c.edit_in_form = false
   end
 end